+++ test/test_core/test_node.py
@@ -7,7 +7,28 @@
 import pytest
 from atraxiflow.core import *
 from atraxiflow.properties import Property
+from atraxiflow.exceptions import PropertyException
+from atraxiflow.base.resources import *
 
+class MakeResources1Node(Node):
+
+    def __init__(self, properties=None):
+        super().__init__({}, properties)
+
+    def run(self, ctx: WorkflowContext) -> bool:
+        self.output.add(TextResource('Text 1'))
+        self.output.add(TextResource('Text 2'))
+        return True
+
+class MakeResources2Node(Node):
+
+    def __init__(self, properties=None):
+        super().__init__({}, properties)
+
+    def run(self, ctx: WorkflowContext) -> bool:
+        self.output.add(TextResource('Text 3'))
+        self.output.add(TextResource('Text 4'))
+        return True
 
 class DemoNode(Node):
     """
@@ -50,6 +71,7 @@ def test_multi_type_property():
 
     test.property('prop_multi').set_value(True)
     test.property('prop_multi').set_value(123)
+
     with pytest.raises(ValueError):
         test.property('prop_multi').set_value([])
 
@@ -64,3 +86,60 @@ def test_fill_properties():
     assert test.property('prop_str').value() == 'Hello World'
     assert test.property('prop_bool').value() == False
     assert test.property('prop_list').value() == ['my_list']
+
+def test_fail_on_unknown_option():
+
+    with pytest.raises(PropertyException):
+        test = DemoNode({
+            'prop_unknown': 'Hello World'
+        })
+
+
+def test_add_to_container():
+
+    c1 = Container()
+    c2 = Container()
+
+    c1.add(TextResource(''))
+    c1.add(TextResource(''))
+    assert c1.size() == 2
+
+    c2.add(TextResource(''))
+    c2.add(TextResource(''))
+    assert c2.size() == 2
+
+    c1.add(c2)
+    assert c1.size() == 4
+
+    with pytest.raises(ValueError):
+        c1.add("Hello World")
+
+def test_passthru_default():
+    
+    n1 = MakeResources1Node()
+    n2 = MakeResources2Node()
+
+    assert n1.output.size() == 0
+    assert n2.output.size() == 0
+    assert not n1.get_passthru()
+
+    assert Workflow.create([n1, n2]).run()
+    assert n1.output.size() == 2
+    assert n2.output.size() == 2
+
+def test_passthru_active():
+    
+    n1 = MakeResources1Node()
+    n2 = MakeResources2Node()
+    n1.set_passthru(True)
+    n2.set_passthru(True)
+
+    assert n1.output.size() == 0
+    assert n2.output.size() == 0
+    assert n2.get_passthru()
+
+    assert Workflow.create([n1, n2]).run()
+    assert n1.output.size() == 2
+    assert n2.output.size() == 4
+
+